monitor/rest_api/serializers.py
from web.models import Agent, Metric, Alert
from rest_framework import serializers


class AgentSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Agent
        fields = "__all__"


class MetricSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Metric
        fields = ["url", "agent", "created", "status", "metrics"]


class AlertSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Alert
        fields = "__all__"
