requirements.txt
asgiref==3.5.1; python_version >= "3.7"
autopep8==1.6.0
backports.zoneinfo==0.2.1; python_version < "3.9" and python_version >= "3.8"
certifi==2021.10.8; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0"
charset-normalizer==2.0.12; python_full_version >= "3.6.0" and python_version >= "3"
click==8.1.3; python_version >= "3.7"
colorama==0.4.4; python_version >= "3.7" and python_full_version < "3.0.0" and platform_system == "Windows" or platform_system == "Windows" and python_version >= "3.7" and python_full_version >= "3.5.0"
defusedxml==0.7.1; python_version >= "3.7" and python_full_version < "3.0.0" or python_version >= "3.7" and python_full_version >= "3.5.0"
diff-match-patch==20200713; python_version >= "3.7"
dj-database-url==0.5.0
django-environ==0.8.1; python_version >= "3.4" and python_version < "4"
django-import-export==2.8.0; python_version >= "3.7"
django==4.0.4; python_version >= "3.8"
djangorestframework==3.13.1; python_version >= "3.6"
et-xmlfile==1.1.0; python_version >= "3.7"
gunicorn==20.1.0; python_version >= "3.5"
h11==0.13.0; python_version >= "3.7"
idna==3.3; python_version >= "3.5" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.5"
markuppy==1.14; python_version >= "3.7"
odfpy==1.4.1; python_version >= "3.7"
openpyxl==3.0.9; python_version >= "3.7"
psycopg2-binary==2.9.3; python_version >= "3.6"
pycodestyle==2.8.0; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.5.0")
pymemcache==3.5.2
pytz==2021.3
pyyaml==6.0; python_version >= "3.7"
requests==2.27.1; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.6.0")
six==1.16.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.3.0"
sqlparse==0.4.2; python_version >= "3.5"
tablib==3.2.1; python_version >= "3.7"
toml==0.10.2; python_version >= "2.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0"
tzdata==2022.1; sys_platform == "win32" and python_version >= "3.8"
unipath==1.1
urllib3==1.26.9; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version < "4"
uvicorn==0.17.6; python_version >= "3.7"
whitenoise==6.0.0; python_version >= "3.7"
xlrd==2.0.1; python_version >= "3.7" and python_full_version < "3.0.0" or python_version >= "3.7" and python_full_version >= "3.6.0"
xlwt==1.3.0; python_version >= "3.7"
